task done

+++ api/urls/urls.controllers.js
@@ -1,23 +1,26 @@
-const Url = require('../../models/Url');
-const shortid = require('shortid');
-const User = require('../../models/User');
+const Url = require("../../models/Url");
+const shortid = require("shortid");
+const User = require("../../models/User");
 
-const baseUrl = 'http:localhost:8000';
+const baseUrl = "http:localhost:8000";
 
 exports.shorten = async (req, res) => {
   // create url code
   const urlCode = shortid.generate();
   try {
-    req.body.shortUrl = baseUrl + '/' + urlCode;
+    req.body.shortUrl = baseUrl + "/" + urlCode;
     req.body.urlCode = urlCode;
-    req.body.userId = req.params.userId;
+    req.body.userId = req.user._id;
+    // req.body.userId = req.user._id;
     const newUrl = await Url.create(req.body);
-    await User.findByIdAndUpdate(req.params.userId, {
+    // console.log("id    " + req.user._id);
+    await User.findByIdAndUpdate(req.user._id, {
       $push: { urls: newUrl._id },
     });
+    // line 16  await User.findByIdAndUpdate(req.params.userId, {
     res.json(newUrl);
   } catch (err) {
-    res.status(500).json('Server Error');
+    res.status(500).json("Server Error");
   }
 };
 
@@ -27,10 +30,10 @@ exports.redirect = async (req, res) => {
     if (url) {
       return res.redirect(url.longUrl);
     } else {
-      return res.status(404).json('No URL Found');
+      return res.status(404).json("No URL Found");
     }
   } catch (err) {
-    res.status(500).json('Server Error');
+    res.status(500).json("Server Error");
   }
 };
 
@@ -38,12 +41,19 @@ exports.deleteUrl = async (req, res) => {
   try {
     const url = await Url.findOne({ urlCode: req.params.code });
     if (url) {
-      await Url.findByIdAndDelete(url._id);
-      return res.status(201).json('Deleted');
+      console.log(
+        "url.userId     " + url.userId + "    req.user._id    " + req.user._id
+      );
+      if (url.userId.toString() == req.user._id.toString()) {
+        await Url.findByIdAndDelete(url._id);
+        return res.status(201).json("Deleted");
+      } else {
+        return res.status(404).json("Not authorized to delete the URL");
+      }
     } else {
-      return res.status(404).json('No URL Found');
+      return res.status(404).json("No URL Found");
     }
   } catch (err) {
-    res.status(500).json('Server Error');
+    res.status(500).json("Server Error");
   }
 };
